remove directory vector_tools
