+++ include/exadg/operators/solution_interpolation_between_triangulations.h
@@ -19,9 +19,10 @@
  *  ______________________________________________________________________
  */
 
-#ifndef INCLUDE_VECTOR_TOOLS_SOLUTION_INTERPOLATION_BETWEEN_TRIANGULATIONS_H
-#define INCLUDE_VECTOR_TOOLS_SOLUTION_INTERPOLATION_BETWEEN_TRIANGULATIONS_H
+#ifndef INCLUDE_EXADG_OPERATORS_SOLUTION_INTERPOLATION_BETWEEN_TRIANGULATIONS_H
+#define INCLUDE_EXADG_OPERATORS_SOLUTION_INTERPOLATION_BETWEEN_TRIANGULATIONS_H
 
+// deal.II
 #include <deal.II/base/mpi_remote_point_evaluation.h>
 #include <deal.II/dofs/dof_handler.h>
 #include <deal.II/fe/mapping.h>
@@ -162,4 +163,4 @@ private:
 
 } // namespace ExaDG
 
-#endif /* INCLUDE_VECTOR_TOOLS_SOLUTION_INTERPOLATION_BETWEEN_TRIANGULATIONS_H */
+#endif /* INCLUDE_EXADG_OPERATORS_SOLUTION_INTERPOLATION_BETWEEN_TRIANGULATIONS_H */
